Minor refactor of DUnitLauncher, DUnitHost, Host
